# docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "1.0.0",
    "title": "Contacts app",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "description": "This is a documentation of contacts app"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    },
    {
      "url": "https://contacts-book-0404/api/v1"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/auth/register": {
      "post": {
        "summary": "Register a new user",
        "operationId": "registerUser",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "tom@gmail.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "123456789"
                  },
                  "name": {
                    "type": "string",
                    "example": "Tom"
                  }
                },
                "required": [
                  "email",
                  "password",
                  "name"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successfully user created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/registationSchema"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Login a user",
        "operationId": "loginUser",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "tom@example.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "1234567"
                  }
                },
                "required": [
                  "email",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully login",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/loginSchema"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "summary": "Logout user",
        "operationId": "logoutUser",
        "tags": [
          "Auth"
        ],
        "parameters": [
          {
            "in": "cookie",
            "name": "refreshToken",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successfully logged out"
          },
          "400": {
            "description": "Bad request, typically missing or invalid refresh token",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 401
                    },
                    "message": {
                      "type": "string",
                      "example": "Unauthorized"
                    },
                    "data": {
                      "type": "string",
                      "example": "\"email\" is required, \"emai\" is not allowed"
                    }
                  },
                  "required": [
                    "status",
                    "message"
                  ]
                },
                "example": {
                  "status": 400,
                  "message": "Invalid refresh token",
                  "data": "Refresh token cookie is missing"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized, if no valid session exists",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 401
                    },
                    "message": {
                      "type": "string",
                      "example": "Unauthorized"
                    },
                    "data": {
                      "type": "string",
                      "example": "\"email\" is required, \"emai\" is not allowed"
                    }
                  },
                  "required": [
                    "status",
                    "message"
                  ]
                },
                "example": {
                  "status": 401,
                  "message": "Something went wrong",
                  "data": "Not authenticated"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "summary": "Refresh the access token",
        "operationId": "refreshToken",
        "tags": [
          "Auth"
        ],
        "parameters": [
          {
            "in": "cookie",
            "name": "refreshToken",
            "description": "The refresh token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "New access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/loginSchema"
                },
                "example": {
                  "status": 200,
                  "message": "Successfully refreshed a session!",
                  "data": {
                    "accessToken": "xi8deii1acfv2gCp06mMjNyDGyrw6pkd+OlOur5s"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 401
                    },
                    "message": {
                      "type": "string",
                      "example": "Unauthorized"
                    },
                    "data": {
                      "type": "string",
                      "example": "\"email\" is required"
                    }
                  },
                  "required": [
                    "status",
                    "message"
                  ]
                },
                "example": {
                  "status": 400,
                  "message": "Invalid refresh token",
                  "data": "Refresh token cookie is missing or malformed"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 401
                    },
                    "message": {
                      "type": "string",
                      "example": "Unauthorized"
                    },
                    "data": {
                      "type": "string",
                      "example": "\"email\" is required, \"emai\" is not allowed"
                    }
                  },
                  "required": [
                    "status",
                    "message"
                  ]
                },
                "example": {
                  "status": 401,
                  "message": "Something went wrong",
                  "data": "Refresh token is invalid"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
    "/auth/reset-pwd": {
      "post": {
        "summary": "Reset user password",
        "operationId": "resetPassword",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": {
                    "type": "string",
                    "description": "JWT token received via email",
                    "example": "twsQHjl84Kg00yf2DNhjjjriqpPlMvbSfj4.eyJ1c2VyX2lkIjoiNjBjNzJiMmY1ZjFiMmMwMDFjOGU4ZGUzZiIsImlhdCI6MTYyMjMwMjU3NywiZXhwIjoxNjIyMzA2MTM3fQ.PVg3zAYnN2vqkpJGbLJebHETK-3Z0dF3Vop5FOgvbbb"
                  },
                  "password": {
                    "type": "string",
                    "description": "to set a new password",
                    "example": "12345678"
                  }
                },
                "required": [
                  "token",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password has been successfully reset",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Password has been successfully reset."
                    },
                    "data": {
                      "type": "object",
                      "example": {}
                    }
                  },
                  "required": [
                    "status",
                    "message",
                    "data"
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "description": "Invalid or expired token",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 401
                    },
                    "message": {
                      "type": "string",
                      "example": "Unauthorized"
                    },
                    "data": {
                      "type": "string",
                      "example": "\"email\" is required, \"emai\" is not allowed"
                    }
                  },
                  "required": [
                    "status",
                    "message"
                  ]
                },
                "example": {
                  "status": 401,
                  "message": "Something went wrong",
                  "data": "Token is expired or invalid."
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "responses": {
      "400": {
        "description": "Invalid request",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 401
                },
                "message": {
                  "type": "string",
                  "example": "Unauthorized"
                },
                "data": {
                  "type": "string",
                  "example": "\"email\" is required, \"emai\" is not allowed"
                }
              },
              "required": [
                "status",
                "message"
              ]
            },
            "example": {
              "status": 400,
              "message": "Something went wrong",
              "data": "\"email\" is required"
            }
          }
        }
      },
      "401": {
        "description": "Unauthorized error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message",
                "data"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 401
                },
                "message": {
                  "type": "string",
                  "example": "UnauthorizedError"
                },
                "data": {
                  "type": "object",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Access token expired"
                    }
                  }
                }
              }
            }
          }
        }
      },
...
      "404": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            },
            "example": {
              "message": "Not found"
            }
          }
        }
      },
      "500": {
        "description": "Internal server error.",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 500
                },
                "message": {
                  "type": "string",
                  "example": "Something went wrong"
                },
                "data": {
                  "type": "string",
                  "example": "Error info"
                }
              }
            },
            "example": {
              "status": 500,
              "message": "Something went wrong",
              "data": "Error info"
            }
          }
        }
      },
...
    },
    "schemas": {
...
      "registationSchema": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "example": 201
          },
          "message": {
            "type": "string",
            "example": "User successfully registered."
          },
          "data": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "example": "67f39f29e3c363521e0ae31d"
              },
              "name": {
                "type": "string",
                "example": "Tom"
              },
              "email": {
                "type": "string",
                "example": "tom@gmail.com"
              }
            }
          }
        },
        "required": [
          "status",
          "message",
          "data"
        ]
      },
      "loginSchema": {
        "type": "object",
        "properties": {
          "status": {
            "type": "number",
            "description": "HTTP status code",
            "example": 200
          },
          "message": {
            "type": "string",
            "example": "Successfully logged in a user!"
          },
          "data": {
            "type": "object",
            "properties": {
              "accessToken": {
                "type": "string",
                "description": "JWT access token",
                "example": "fggvhjrsxbmk/8yJu3Bin3W6ff6f3kkZHcVT+J+jdf"
              }
            }
          }
        },
        "required": [
          "status",
          "message",
          "data"
        ]
      }
    }
  }
}